Add a new "trioisms" test mod for tracking `trio` wishlist behaviour
