+++ tests/test_child_manages_service_nursery.py
@@ -171,60 +171,3 @@ def test_actor_managed_trio_nursery_task_error_cancels_aio(
     # verify boxed error
     err = excinfo.value
     assert isinstance(err.type(), NameError)
-
-
-def test_stashed_child_nursery():
-
-    _child_nursery = None
-
-    async def waits_on_signal(
-        ev: trio.Event(),
-        task_status: TaskStatus[trio.Nursery] = trio.TASK_STATUS_IGNORED,
-    ):
-        '''
-        Do some stuf, then signal other tasks, then yield back to "starter".
-
-        '''
-        await ev.wait()
-        task_status.started()
-
-    async def mk_child_nursery(
-        task_status: TaskStatus = trio.TASK_STATUS_IGNORED,
-    ):
-        '''
-        Allocate a child sub-nursery and stash it as a global.
-
-        '''
-        nonlocal _child_nursery
-
-        async with trio.open_nursery() as cn:
-            _child_nursery = cn
-            task_status.started(cn)
-
-            # block until cancelled by parent.
-            await trio.sleep_forever()
-
-    async def sleep_and_err(ev: trio.Event):
-        await trio.sleep(0.5)
-        doggy()  # noqa
-        ev.set()
-
-    async def main():
-
-        async with (
-            trio.open_nursery() as pn,
-        ):
-            cn = await pn.start(mk_child_nursery)
-            assert cn
-
-            ev = trio.Event()
-            cn.start_soon(sleep_and_err, ev)
-
-            # this causes inf hang
-            await cn.start(waits_on_signal, ev)
-
-            # this does not.
-            # cn.start_soon(waits_on_signal, ev)
-
-    with pytest.raises(NameError):
-        trio.run(main)
